모집공고문PDF추출/requirements.txt
google-genai>=0.3.0
pydantic>=2.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
